feat(vaccine): 新增疫苗记录并创建定时任务
- 新增疫苗记录时，如果需要提醒，则创建对应的定时任务
- 定义了任务类型的枚举和有效性验证
- 优化了参数绑定和验证逻辑
- 增加了事务处理，确保数据一致性

+++ middleware/token_middleware.go
@@ -45,7 +45,7 @@ func AuthMiddleware() gin.HandlerFunc {
 
 		// 提取用户信息
 		if claims, ok := token.Claims.(*model.UserStdClaim); ok && token.Valid {
-			c.Set("userId", claims.Id)         // 将 userID 存入上下文
+			c.Set("userId", uint(claims.Id))   // 将 userID 存入上下文
 			c.Set("nickName", claims.NickName) // 将 nickName 存入上下文
 		} else {
 			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "Invalid token claims"})
